Extract NovaCache.Server: the HTTP protocol as brand-free library API
Routing, write auth, body limits, and the narinfo validation/signing
pipeline move into the library; the executable keeps configuration, key
loading, and its own landing page (embedders supply theirs).  Hardening
riding along: /narinfo-hashes is authenticated and no-store, HEAD is
served wherever GET is, NAR transfer streams via the new FileStore
primitives, an empty-after-normalize CACHE_API_KEY refuses startup, and
the bind host is configurable (--host / HOST, default unchanged).
