+++ nova-cache.cabal
@@ -35,6 +35,7 @@ library
     NovaCache.Hash
     NovaCache.NAR
     NovaCache.NarInfo
+    NovaCache.Server
     NovaCache.Signing
     NovaCache.Store
     NovaCache.StorePath
@@ -48,9 +49,11 @@ library
     , crypton             >= 1.1 && < 2
     , directory           >= 1.3 && < 1.4
     , filepath            >= 1.4 && < 1.6
+    , http-types          >= 0.12 && < 0.13
     , ram                 >= 0.20 && < 1
     , text                >= 2.0 && < 2.2
     , vector              >= 0.12 && < 0.14
+    , wai                 >= 3.2 && < 3.3
 
 
   hs-source-dirs:   src
@@ -69,6 +72,7 @@ executable nova-cache-server
     buildable: False
 
   main-is:          Main.hs
+  other-modules:    LandingPage
   hs-source-dirs:   exe
   default-language:  Haskell2010
   default-extensions:
@@ -79,7 +83,6 @@ executable nova-cache-server
       base                >= 4.16 && < 5
     , bytestring          >= 0.11 && < 0.13
     , nova-cache
-    , ram                 >= 0.20 && < 1
     , http-types          >= 0.12 && < 0.13
     , text                >= 2.0 && < 2.2
     , wai                 >= 3.2 && < 3.3
@@ -101,9 +104,12 @@ test-suite nova-cache-test
     , bytestring          >= 0.11 && < 0.13
     , crypton             >= 1.1 && < 2
     , directory           >= 1.3 && < 1.4
+    , http-types          >= 0.12 && < 0.13
     , nova-cache
     , ram                 >= 0.20 && < 1
     , text                >= 2.0 && < 2.2
+    , wai                 >= 3.2 && < 3.3
+    , wai-extra           >= 3.1 && < 3.2
 
 source-repository head
   type:     git
